some updates for C
